feat: 1.1.0 - Updated libraries, added more tweaks & some internal changes

+++ src/main/kotlin/dev/tarna/moretweaks/tweaks/BatMembranes.kt
@@ -26,7 +26,7 @@ class BatMembranes : Tweak {
     fun onBatDeath(event: EntityDeathEvent) {
         val entity = event.entity
         if (entity !is Bat) return
-        if (!chance(chance.toInt())) return
+        if (!chance(chance)) return
 
         val location = entity.location
         location.world.dropItemNaturally(location, drop.toItemStack(amount.toInt()))
